module_27_postgres_migrations/hw/task_1/pyproject.toml
[tool.poetry]
name = "fast-api-with-postgres"
version = "0.1.0"
description = "docker images"
authors = ["Miron <mairon26rus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
